reducing code smells
